fixed memory leak in Quadtree

+++ src/athi_settings.cpp
@@ -23,8 +23,8 @@ bool show_fps_info{true};
 
 bool quadtree_show_only_occupied{true};
 bool quadtree_active{false};
-int quadtree_depth{5};
-int quadtree_capacity{50};
+size_t quadtree_depth{5};
+size_t quadtree_capacity{50};
 
 bool voxelgrid_active{false};
 int voxelgrid_parts{4};
